requirementss.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangrestframework==0.0.1
environs==11.0.0
idna==3.6
marshmallow==3.21.1
packaging==24.0
peppercorn==0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
